# src/main/resources/openAPI/TransactionAPI.yaml
openapi: 3.0.2
info:
  description: Provides the The Nordic Smart Government's transaction collection in XBRL GL
  version: '0.1'
  title: The Nordic Smart Government's transaction collection API
  contact:
    name: Nordic Smart Government
    url: 'https://nordicsmartgovernment.github.io'
  license:
    name: CC0 1.0 Universal (CC0 1.0)
    url: 'http://creativecommons.org/publicdomain/zero/1.0/deed.no'
externalDocs:
  description: Documentation on nordicsmartgovernment docs
  url: https://nordicsmartgovernment.github.io
servers:
  - url: 'https://api.nordicsmartgovernment.org'
    description: The reference implementation server
tags:
  - name: transactionCollection
    description: transactionCollection operations
  - name: writeTransaction
    description: operations that writes or updates a single transcation
paths:
  /transactions/{companyId}:
    get:
      tags:
        - transactionCollection
      description: List every transaction in the catalog
      operationId: getTransactions
      parameters:
        - name: companyId
          in: path
          description: The transactions owner customerId
          required: true
          schema:
            type: string
        - name: documentId
          in: query
          description: Filter transactions by document id
          required: false
          schema:
            type: string
        - name: invoiceType
          in: query
          description: Filter transactions by the type of the invoice
          required: false
          schema:
            type: string
            enum:
            - incoming
            - outgoing
            - all
            default: all
      responses:
        '200':
          description: OK
          content:
            application/xbrl-instance+xml:
              schema:
                type: object
            application/json:
              schema:
                type: object
  /transactions/{companyId}/{id}:
    get:
      tags:
        - transactionCollection
      description: Get a transaction by transaction id
      operationId: getTransactionById
      parameters:
        - name: companyId
          in: path
          description: The transaction owner customerId
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          description: The id of the Transaction
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/xbrl-instance+xml:
              schema:
                $ref: "#/components/schemas/Transaction"
    put:
      tags:
        - writeTransaction
      description: Patch a transaction
      operationId: putTransactionById
      parameters:
        - name: companyId
          in: path
          description: The transaction owner customerId
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          description: The id of the transaction
          schema:
            type: string
      responses:
        '200':
          description: OK
      requestBody:
        required: true
        content:
          application/xbrl-instance+xml:
            schema:
              type: string
components:
  schemas:
    Transaction:
      type: object